+++ packages/blitz/package.json
@@ -29,7 +29,6 @@
     "watch": "1.0.2"
   },
   "dependencies": {
-    "cookie": "0.4.1",
     "chalk": "4.1.0",
     "console-table-printer": "2.10.0",
     "ora": "5.3.0",
